Fix crash on iOS
It looks like UIImage initializers have become non thread-safe.
Altought the current code could still enter in race conditions
with the app code, at least now it won't interfere with itself.

If we find this to be a problem in the future, code for instantiating
UIImage objects will be moved to the main thread. Since the decoding
is already done using ImageIO in the background, the performance
penalty should not be noticeable.